refactor: Update styles on error message
